add conditional for drop table

+++ db/migrate/20140529225707_drop_bookpages_table.rb
@@ -1,5 +1,5 @@
 class DropBookpagesTable < ActiveRecord::Migration
   def change
-  	drop_table :bookpages
+  	drop_table :bookpages if column_exists?(:bookpages, :title)
   end
 end
